Add files via upload
This program/project accepts an XLSX file that contains a series of employees/persons info for a specific company or institution and outputs the data in an easy-to-read XML file format after taking a user command if sorting is needed regarding first name, working years, salary, or gender. The program is built using Java and the use of Apache package to read the cells of the workbook page in the XLSX file.